[issue-17] extract hexcode colors into notris_web
