+++ apps/notris/lib/notris/color.ex
@@ -7,8 +7,6 @@ defmodule Notris.Color do
 
   @type t :: :red | :green | :blue | :yellow | :orange | :purple | :grey
 
-  @type hexcode :: String.t()
-
   @shape_color_map %{
     i: :red,
     l: :green,
@@ -17,17 +15,6 @@ defmodule Notris.Color do
     s: :orange,
     t: :purple,
     z: :grey
-  }
-
-  # https://htmlcolorcodes.com/color-chart/
-  @hexcodes %{
-    red: {"#CE351B", "#A50C00"},
-    green: {"#34B515", "#34B515"},
-    blue: {"#049CBF", "#047BA0"},
-    yellow: {"#F9E79F", "#D4AC0D"},
-    orange: {"#D66203", "#EF880C"},
-    purple: {"#835EA5", "#683C96"},
-    grey: {"#CCD1D1", "#707B7C"}
   }
 
   @values Map.values(@shape_color_map)
@@ -42,10 +29,5 @@ defmodule Notris.Color do
   @spec color_of(Shape.t()) :: t()
   def color_of(shape) do
     @shape_color_map[shape]
-  end
-
-  @spec hexcodes_of(t()) :: hexcode()
-  def hexcodes_of(color) do
-    @hexcodes[color]
   end
 end
